# feature: use Rosetta AOT Caching with CDI