@norio-nomura 2025-08-20
Renamed `lima.io/rosetta=cached` to `lima-vm.io/rosetta=cached`.
Also edited examples in the above comments.