@AkihiroSuda 2025-08-20
Let's mark this as a draft until https://github.com/Code-Hex/vz/pull/195 is merged